Added WHO scraping method, modified JSON saving
WHO scraping method seems to work on my end, uses two new libraries, selenium and webdriver-manager. Make sure to install these before running! WHO method is pretty jank, but it works. Basically, opens a chrome instance of the WHO page and clicks the table filter button a bunch to get the output right THEN passes the website data to beautifulsoup so that the correct columns are being scraped.  Also modified the JSON saving method to concat on to a existing JSON if one exists, based on actual dates instead of "today", "yesterday".